Splitup Cassandra integration test collections.rs
