+++ shotover-proxy/tests/cassandra_int_tests/collections/mod.rs
@@ -4,6 +4,10 @@ use test_helpers::connection::cassandra::{
     assert_query_result, run_query, CassandraConnection, CassandraDriver, ResultValue,
 };
 
+mod list;
+mod map;
+mod set;
+
 const NATIVE_COL_TYPES: [ColType; 18] = [
     ColType::Ascii,
     ColType::Bigint,
@@ -294,317 +298,6 @@ async fn select_table(
         &[&results],
     )
     .await;
-}
-
-mod list {
-    use super::*;
-
-    async fn create(connection: &CassandraConnection) {
-        create_table(connection, "list_native", |s| format!("list<{s}>")).await;
-        create_table(connection, "list_set", |s| {
-            format!("list<frozen<set<{s}>>>")
-        })
-        .await;
-        create_table(connection, "list_list", |s| {
-            format!("list<frozen<list<{s}>>>")
-        })
-        .await;
-        create_table(connection, "list_map", |s| {
-            format!("list<frozen<map<int, {s}>>>")
-        })
-        .await;
-    }
-
-    async fn insert(connection: &CassandraConnection) {
-        insert_table(connection, "list_native", |elements| {
-            format!("[{}]", elements.join(","))
-        })
-        .await;
-        insert_table(connection, "list_set", |elements| {
-            let set: String = format!("{{{}}}", elements.join(","));
-            format!("[{}]", (0..3).map(|_| &set).join(","))
-        })
-        .await;
-        insert_table(connection, "list_list", |elements| {
-            let list = format!("[{}]", elements.join(","));
-            format!("[{}]", (0..3).map(|_| &list).join(","))
-        })
-        .await;
-        insert_table(connection, "list_map", |elements| {
-            let map = format!(
-                "{{{}}}",
-                elements
-                    .iter()
-                    .enumerate()
-                    .map(|(i, element)| { format!("{i}: {element}") })
-                    .join(",")
-            );
-
-            format!("[{}]", (0..3).map(|_| &map).join(","))
-        })
-        .await;
-    }
-
-    async fn select(connection: &CassandraConnection, driver: CassandraDriver) {
-        select_table(connection, "list_native", ResultValue::List, driver).await;
-        select_table(
-            connection,
-            "list_set",
-            |t| {
-                ResultValue::List(vec![
-                    ResultValue::Set(t.clone()),
-                    ResultValue::Set(t.clone()),
-                    ResultValue::Set(t),
-                ])
-            },
-            driver,
-        )
-        .await;
-        select_table(
-            connection,
-            "list_list",
-            |t| {
-                ResultValue::List(vec![
-                    ResultValue::List(t.clone()),
-                    ResultValue::List(t.clone()),
-                    ResultValue::List(t),
-                ])
-            },
-            driver,
-        )
-        .await;
-        select_table(
-            connection,
-            "list_map",
-            |t| {
-                let map = ResultValue::Map(
-                    t.iter()
-                        .enumerate()
-                        .map(|(i, x)| (ResultValue::Int(i as i32), x.clone()))
-                        .collect(),
-                );
-                ResultValue::List(vec![map.clone(), map.clone(), map])
-            },
-            driver,
-        )
-        .await;
-    }
-
-    pub async fn test(connection: &CassandraConnection, driver: CassandraDriver) {
-        create(connection).await;
-        insert(connection).await;
-        select(connection, driver).await;
-    }
-}
-
-mod set {
-    use super::*;
-
-    async fn create(connection: &CassandraConnection) {
-        create_table(connection, "set_native", |s| format!("set<{s}>")).await;
-        create_table(connection, "set_set", |s| format!("set<frozen<set<{s}>>>")).await;
-        create_table(connection, "set_list", |s| {
-            format!("set<frozen<list<{s}>>>")
-        })
-        .await;
-        create_table(connection, "set_map", |s| {
-            format!("set<frozen<map<int, {s}>>>")
-        })
-        .await;
-    }
-
-    async fn insert(connection: &CassandraConnection) {
-        insert_table(connection, "set_native", |elements| {
-            format!("{{{}}}", elements.join(","))
-        })
-        .await;
-        insert_table(connection, "set_set", |elements| {
-            let set = format!("{{{}}}", elements.join(","));
-            format!("{{{}}}", (0..3).map(|_| &set).join(","))
-        })
-        .await;
-        insert_table(connection, "set_list", |elements| {
-            let list = format!("[{}]", elements.join(","));
-            format!("{{{}}}", (0..3).map(|_| &list).join(","))
-        })
-        .await;
-        insert_table(connection, "set_map", |elements| {
-            let map = format!(
-                "{{{}}}",
-                elements
-                    .iter()
-                    .enumerate()
-                    .map(|(i, element)| { format!("{i}: {element}") })
-                    .join(",")
-            );
-
-            format!("{{{}}}", (0..3).map(|_| &map).join(","))
-        })
-        .await;
-    }
-
-    async fn select(connection: &CassandraConnection, driver: CassandraDriver) {
-        select_table(connection, "set_native", ResultValue::Set, driver).await;
-        select_table(
-            connection,
-            "set_set",
-            |t| ResultValue::Set(vec![ResultValue::Set(t.clone())]),
-            driver,
-        )
-        .await;
-        select_table(
-            connection,
-            "set_list",
-            |t| ResultValue::Set(vec![ResultValue::List(t)]),
-            driver,
-        )
-        .await;
-        select_table(
-            connection,
-            "set_map",
-            |t| {
-                ResultValue::Set(vec![ResultValue::Map(
-                    t.iter()
-                        .enumerate()
-                        .map(|(i, x)| (ResultValue::Int(i as i32), x.clone()))
-                        .collect(),
-                )])
-            },
-            driver,
-        )
-        .await;
-    }
-
-    pub async fn test(connection: &CassandraConnection, driver: CassandraDriver) {
-        create(connection).await;
-        insert(connection).await;
-        select(connection, driver).await;
-    }
-}
-
-mod map {
-    use super::*;
-
-    async fn create(connection: &CassandraConnection) {
-        create_table(connection, "map_native", |s| format!("map<int, {s}>")).await;
-        create_table(connection, "map_set", |s| {
-            format!("map<int, frozen<set<{s}>>>")
-        })
-        .await;
-        create_table(connection, "map_list", |s| {
-            format!("map<int, frozen<list<{s}>>>")
-        })
-        .await;
-        create_table(connection, "map_map", |s| {
-            format!("map<int, frozen<map<int, {s}>>>")
-        })
-        .await;
-    }
-
-    async fn insert(connection: &CassandraConnection) {
-        insert_table(connection, "map_native", |elements| {
-            format!(
-                "{{{}}}",
-                elements
-                    .iter()
-                    .enumerate()
-                    .map(|(i, element)| format!("{i}: {element}"))
-                    .join(",")
-            )
-        })
-        .await;
-        insert_table(connection, "map_set", |elements| {
-            let set = format!("{{{}}}", elements.join(","));
-            format!("{{{}}}", (0..3).map(|i| format!("{i}: {set}")).join(","))
-        })
-        .await;
-        insert_table(connection, "map_list", |elements| {
-            let list = format!("[{}]", elements.join(","));
-            format!("{{{}}}", (0..3).map(|i| format!("{i}: {list}")).join(","))
-        })
-        .await;
-        insert_table(connection, "map_map", |elements| {
-            let map = format!(
-                "{{{}}}",
-                elements
-                    .iter()
-                    .enumerate()
-                    .map(|(i, element)| { format!("{i}: {element}") })
-                    .join(",")
-            );
-
-            format!("{{{}}}", (0..3).map(|i| format!("{i}: {map}")).join(","))
-        })
-        .await;
-    }
-
-    async fn select(connection: &CassandraConnection, driver: CassandraDriver) {
-        select_table(
-            connection,
-            "map_native",
-            |t| {
-                ResultValue::Map(
-                    t.iter()
-                        .enumerate()
-                        .map(|(i, x)| (ResultValue::Int(i as i32), x.clone()))
-                        .collect(),
-                )
-            },
-            driver,
-        )
-        .await;
-        select_table(
-            connection,
-            "map_set",
-            |t| {
-                ResultValue::Map(vec![
-                    (ResultValue::Int(0), ResultValue::Set(t.clone())),
-                    (ResultValue::Int(1), ResultValue::Set(t.clone())),
-                    (ResultValue::Int(2), ResultValue::Set(t)),
-                ])
-            },
-            driver,
-        )
-        .await;
-        select_table(
-            connection,
-            "map_list",
-            |t| {
-                ResultValue::Map(vec![
-                    (ResultValue::Int(0), ResultValue::List(t.clone())),
-                    (ResultValue::Int(1), ResultValue::List(t.clone())),
-                    (ResultValue::Int(2), ResultValue::List(t)),
-                ])
-            },
-            driver,
-        )
-        .await;
-        select_table(
-            connection,
-            "map_map",
-            |t| {
-                let map = ResultValue::Map(
-                    t.iter()
-                        .enumerate()
-                        .map(|(i, x)| (ResultValue::Int(i as i32), x.clone()))
-                        .collect(),
-                );
-                ResultValue::Map(vec![
-                    (ResultValue::Int(0), map.clone()),
-                    (ResultValue::Int(1), map.clone()),
-                    (ResultValue::Int(2), map),
-                ])
-            },
-            driver,
-        )
-        .await;
-    }
-
-    pub async fn test(connection: &CassandraConnection, driver: CassandraDriver) {
-        create(connection).await;
-        insert(connection).await;
-        select(connection, driver).await;
-    }
 }
 
 pub async fn test(connection: &CassandraConnection, driver: CassandraDriver) {
